LayoutTests/discovery/video-player-has-text-track-expected.txt

Tests HTTPMediaElement Player.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Text Track added.
PASS vid.textTracks.length>0 is true
PASS vid.textTracks.length>0 is true
PASS successfullyParsed is true

TEST COMPLETE
